% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candLoc.R
\name{cand_loc_searcher}
\alias{cand_loc_searcher}
\title{Candidate Location Searcher}
\usage{
cand_loc_searcher(
  gtfs.obj,
  num.stops = 10,
  syn.network = NULL,
  all.routes = TRUE,
  ag.name = NULL,
  opacity = 0.7,
  verbose = TRUE,
  include.legend = TRUE,
  tileset = NULL,
  include.depots = TRUE
)
}
\arguments{
\item{gtfs.obj}{a GTFS object, preferably generated by the function gtfs_loader}

\item{num.stops}{the number of stops to display on the output. This value is not used if \code{all.routes = TRUE}}

\item{syn.network}{a vector of routes as a synthetic network, or NULL if the entire network is to be used}

\item{all.routes}{a boolean. If set to \code{TRUE}, candidate locations are selected such that every route has at least one location to charge at (this ignores \code{num.stops}). If \code{FALSE}, only the top locations are selected regardless of if routes are left out.}

\item{ag.name}{character, the title of the legend in the map. Optional; if blank, the information will be pulled from \code{agency}.}

\item{opacity}{numeric, the opacity value of the routes (higher numbers are darker)}

\item{verbose}{a boolean. If \code{TRUE}, diagnostic information will be displayed.}

\item{include.legend}{a boolean. If \code{TRUE}, a legend for routes will be displayed on the map output.}

\item{tileset}{a character naming the \code{leaflet} tileset to use. See \code{names(providers)} with leaflet attached for a list of values. If \code{NULL}, the default leaflet tileset will be used.}

\item{include.depots}{a boolean. If \code{TRUE}, depots will be included as red pins if depot data is available.}
}
\value{
a list containing a histogram of the stops, a table of stops, a map of the routes and stops, and the list of stops at each location.
}
\description{
This function returns the top number of times a stop is stopped at in a network. It should be noted that the total number of times a bus is scheduled to stop at a bus is all that is considered.
}
