% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dater}
\alias{dater}
\title{Convert time to dated time}
\usage{
dater(time)
}
\arguments{
\item{time}{a character representing a time in the form \%H:\%M:\%S. Hour values over 24 are acceptable.}
}
\value{
POSIXct
}
\description{
Converts the inputted time character to a POSIXct dated time. This is primarily used to handle schedules that run past midnight.
}
\examples{
dater("14:35:43")
}
