% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyTabler.R
\name{day_tabler}
\alias{day_tabler}
\title{Day Tabler}
\usage{
day_tabler(calendar = NA, trips = NA)
}
\arguments{
\item{calendar}{data.frame, the output from year_tabler should be used.}

\item{trips}{data.frame, containing trip information generated by gtfs_framer}
}
\value{
list, a list of days for which all trips are included
}
\description{
This function takes the calendar generated by year_tabler and the trip information generated by gtfs_framer and makes a list of all trips for each day of the year.
}
