% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyTabler.R
\name{energy_tabler}
\alias{energy_tabler}
\title{Create Energy Table}
\usage{
energy_tabler(
  gtfs.obj,
  large.Ag,
  mass.vec = NULL,
  mass.labels = NULL,
  input.loc = "./GTFS Local Data/GMapsResults",
  output.loc = "./Energy Tables",
  summary.loc = "./Results/Pre Fuel",
  ignore.Existing = FALSE,
  surpress.output = TRUE,
  save.output = TRUE
)
}
\arguments{
\item{gtfs.obj}{a gtfs object generated by \code{tidytransit::read_gtfs()}.}

\item{large.Ag}{a boolean. If \code{TRUE}, regression will be against large, urban networks.  If \code{FALSE}, regression will be against small, rural networks}

\item{mass.vec}{a vector of masses to calculate energy for. If left as \code{NULL} (default), values for average 35ft, 40ft, and 60ft buses will be calculated.}

\item{mass.labels}{a vector of characters containing the labels for the bus types.  Must match the \code{length()} of \code{mass.vec}}

\item{input.loc}{a character containing the location of the google maps results as an RDS file.}

\item{output.loc}{a character containing the location to save the energy tables.}

\item{summary.loc}{a character containing the location of the summary file generated by \code{gtfs_summer}}

\item{surpress.output}{a boolean. If \code{TRUE}, the energy table will not be returned; the function will instead return \code{0} to signify successful completion.  If \code{FALSE}, the energy table is returned.}

\item{save.output}{a boolean. If \code{TRUE}, the output will be saved in the specified \code{output.loc}.}
}
\value{
0 if the table was created successfully (or already exists), or the created energy table (see \code{surpress.output} for more information).
}
\description{
This function takes a GTFS object generated by gtfs_lengthr and develops and saves an energy table for each route in that network.
}
