% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_routes_sldf}
\alias{get_routes_sldf}
\title{Get shapes spatial data for given route ids
This function was originally part of the \code{gtfsr} package; the code was written by its authors:}
\usage{
get_routes_sldf(
  gtfs.obj,
  route_ids,
  service_ids,
  shape_ids,
  route_opacity,
  route_colors
)
}
\arguments{
\item{gtfs.obj}{A GTFS list object with components agency, etc.}

\item{route_ids}{Vector (Character). IDs for routes of interest.}

\item{service_ids}{Vector (Character). Service IDs. NULL by Default.}

\item{shape_ids}{Vector (Character). Shape IDs. NULL by Default.}

\item{route_opacity}{Numeric. Value must be between 0 and 1. Default is NULL.}

\item{route_colors}{Character. Names of colors (e.g. "blue") or hex values (e.g. '#000000').}
}
\value{
Environment containing spatial data, labels, colorings used for plotting
}
\description{
Elaine McVey (elaine.mcvey\@transloc.com)
Danton Noriega-Goodwin (danton.noriega\@gmail.com)
}
\details{
gtfsr was distributed under the GPL license.  For more information, please see the GPL license website: https://www.gnu.org/licenses/gpl-3.0.en.html
}
