% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyTabler.R
\name{gtfs_framer}
\alias{gtfs_framer}
\title{GTFS Framer}
\usage{
gtfs_framer(gtfs.obj)
}
\arguments{
\item{gtfs.obj}{a gtfs object that has been treated with gtfs_lengthr}
}
\value{
The pertinent information used in other functions to develop the energy table, disentangled into a single data.frame.
}
\description{
This function takes a GTFS Object that has been downloaded with gtfs_loader and treated with gtfs_lengthr to disentangles the trips, routes, and service IDs into a single data.frame.
}
