% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyTabler.R
\name{gtfs_summarize}
\alias{gtfs_summarize}
\title{GTFS Summarize}
\usage{
gtfs_summarize(
  time.table,
  gtfs.obj,
  start.day = 1,
  end.day = 365,
  location = "./Results/Pre Fuel/"
)
}
\arguments{
\item{time.table}{data.frame, generated by day_tabler.}

\item{gtfs.obj}{a gtfs object treated with gtfs_lengthr}

\item{start.day}{integer, the starting day for setup}

\item{end.day}{integer, the ending day for setup}

\item{location}{character string, the location for saving the RDS files. Defaults to "./Results/Pre Fuel/". If the directory doesn't exist, it will be created.}
}
\description{
This function takes a time table generated by day_tabler and a GTFS object treated by gtfs_lengthr and uses those to generate and save a final data.frame and RDS file to be used for energy modeling.
}
