% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jan_first}
\alias{jan_first}
\title{Generate a date for January 1 of the current year}
\usage{
jan_first()
}
\value{
Date, a date for January 1st of the current year
}
\description{
Generate a date for January 1 of the current year
}
