% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{latlon_2_meter}
\alias{latlon_2_meter}
\title{Find Distance between Two Lat/Long Pairs}
\usage{
latlon_2_meter(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{numeric, a latitude}

\item{lon1}{numeric, a longitude}

\item{lat2}{numeric, another latitude}

\item{lon2}{numeric, another longitude}
}
\value{
Numeric, the distance in meters between the points.  Error is untested at this point.
}
\description{
Find Distance between Two Lat/Long Pairs
}
