% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyTabler.R
\name{lengthr_local}
\alias{lengthr_local}
\title{Lengthr Local}
\usage{
lengthr_local(gtfs.obj, local.file = "./GTFS Local Data/GMapsResults")
}
\arguments{
\item{gtfs.obj}{a GTFS object to be lengthed.}

\item{local.file}{the location of the file containing the \code{gmapsdistance} results. Defaults to the same as the output files of \code{gtfs_lengthr}.}
}
\value{
a gtfs object with a modified \code{stop_times_df}, equivalent to the output of \code{gtfs_lengthr}.
}
\description{
Identical functionality to \code{gtfs_lengthr}, except using a local file instead of generating the results using \code{gmapsdistance}.
}
\examples{
lengthr_local(gtfs.obj)
#works as called from gtfs_lengthr

}
