% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rectify_colnames}
\alias{rectify_colnames}
\title{Rectify Column Names}
\usage{
rectify_colnames(gtfs.obj)
}
\arguments{
\item{gtfs.obj}{an object created by TODO}
}
\value{
The GTFS obejct with rectified column names
}
\description{
This function rectifies column names by removing erroneous byte order marks that can sometimes appear when importing GTFS objects.
}
\examples{
rectify_colnames(gtfs.obj)
#Will return the GTFS object supplied, with any byte order makes (commonly appearing in text as "ï..").

}
