% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_on_routes}
\alias{stop_on_routes}
\title{What routes pass through this stop?}
\usage{
stop_on_routes(gtfs.obj, stop.id)
}
\arguments{
\item{gtfs.obj}{a network to analyze.}

\item{stop.id}{a stop id within the network to check on.}
}
\value{
a list of routes that pass through the stop.
}
\description{
What routes pass through this stop?
}
