% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyTabler.R
\name{year_tabler}
\alias{year_tabler}
\title{Year Tabler}
\usage{
year_tabler(gtfs.obj)
}
\arguments{
\item{gtfs.obj}{a gtfs object treated with gtfs_lengthr}
}
\value{
data.frame, a base calendar to use as an input toe day_tabler.
}
\description{
This function builds a base calendar of the year for use in constructing the timetable of a transit network.
}
